chore: extract shared seed skill value builder
Extract duplicated skill seeding logic from dev-seed.ts and
api-test-helpers.ts into a shared buildSeedSkillValues() function
in seed-skills.ts. Migrate dev-seed.ts from standalone postgres
connection to initServices(). Unify versionHash to null in both
contexts.

Closes #6308
