+++ turbo/apps/web/src/__tests__/api-test-helpers.ts
@@ -3437,21 +3437,11 @@ export async function seedTestSkill(
  * succeeds when buildComposeContent injects them.
  */
 export async function seedSeedSkills(): Promise<void> {
-  const { SEED_SKILLS } = await import("../lib/zero/seed-skills");
+  const { SEED_SKILLS, buildSeedSkillValues } = await import(
+    "../lib/zero/seed-skills"
+  );
   initServices();
-  const values = SEED_SKILLS.map((name) => ({
-    url: `https://github.com/vm0-ai/vm0-skills/tree/main/${name}`,
-    name,
-    fullPath: `vm0-ai/vm0-skills/tree/main/${name}`,
-    versionHash:
-      "a1b2c3d4e5f6a7b8c9d0e1f2a3b4c5d6e7f8a9b0c1d2e3f4a5b6c7d8e9f0a1b2",
-    frontmatter: {
-      name,
-      description: `${name} skill`,
-      vm0_secrets: [] as string[],
-      vm0_vars: [] as string[],
-    },
-  }));
+  const values = buildSeedSkillValues(SEED_SKILLS);
   await globalThis.services.db
     .insert(skills)
     .values(values)
